Add partial support for reading resources
This doesn't work properly. As far as I can tell it's reading the
resource correctly, but it isn't connecting to the opening application
properly. I hadn't really intended on this being a use case though
(I was thinking more along the lines of simple editing), so I'm going
to drop this for now. I might come back to it with more motivation
later.