re-work tags
Tags are not included in events, so the behaviour is tweaked so that
tag data is always fetched fresh when needed.

This also fixes a bug with notebooks where all notebooks were being
displayed in the root.
This also fixes a bug where some notebooks were not getting children
added properly. This was caused by the mutations to meta, basically,
the meta from the outer-loop was different from the meta in the
inner-loop (in the construct map function). Depending on which
initiated the cache would determine if the children were established
correctly.